The long awaited install script is here!
It is a Windows Powershell install script! Anyone running on Linux is presumably already knowledable enough of command line to not need a script and I can't make anything for MacOS since I don't have a MacOS machine (pending donation) to test with. As for powershell over batch.. it's 2018. that's why.